feat: simplify greatly
